# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Credit Card Fraud Detection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: status
          in: query
          schema: { type: string, enum: [approved, declined, flagged, pending] }
        - name: riskLevel
          in: query
          schema: { type: string, enum: [low, medium, high, critical] }
        - name: cardId
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"

    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Submit a new transaction for fraud analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Transaction analyzed and created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{id}/review:
    patch:
      operationId: reviewTransaction
      tags: [transactions]
      summary: Approve or decline a flagged transaction
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionReview"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Transaction:
      type: object
      required: [id, amount, merchant, merchantCategory, cardId, status, riskScore, riskLevel, fraudProbability, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        merchant:
          type: string
        merchantCategory:
          type: string
        cardId:
          type: string
        cardLast4:
          type: string
        userId:
          type: integer
        status:
          type: string
          enum: [approved, declined, flagged, pending]
        riskScore:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        fraudProbability:
          type: number
        location:
          type: string
        ipAddress:
          type: string
        deviceId:
          type: string
        reviewNote:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [amount, merchant, merchantCategory, cardId]
      properties:
        amount:
          type: number
        merchant:
          type: string
        merchantCategory:
          type: string
        cardId:
          type: string
        location:
          type: string
        ipAddress:
          type: string
        deviceId:
          type: string
    TransactionReview:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, declined]
        note:
          type: string
    TransactionListResponse:
      type: object
      required: [transactions, total, page, limit]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...