# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionRiskLevel } from './transactionRiskLevel';
import type { TransactionStatus } from './transactionStatus';

export interface Transaction {
  id: number;
  amount: number;
  merchant: string;
  merchantCategory: string;
  cardId: string;
  cardLast4?: string;
  userId?: number;
  status: TransactionStatus;
  riskScore: number;
  riskLevel: TransactionRiskLevel;
  fraudProbability: number;
  location?: string;
  ipAddress?: string;
  deviceId?: string;
  /** @nullable */
  reviewNote?: string | null;
  /** @nullable */
  reviewedBy?: number | null;
  createdAt: string;
}