# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Credit Card Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionInput {
  amount: number;
  merchant: string;
  merchantCategory: string;
  cardId: string;
  location?: string;
  ipAddress?: string;
  deviceId?: string;
}